+++ satellite/src/main/java/com/linbit/linstor/core/objects/EbsRemoteSatelliteFactory.java
@@ -3,13 +3,14 @@ package com.linbit.linstor.core.objects;
 import com.linbit.ImplementationError;
 import com.linbit.linstor.core.CoreModule;
 import com.linbit.linstor.core.CoreModule.RemoteMap;
-import com.linbit.linstor.core.DivergentUuidsException;
+import com.linbit.linstor.core.CriticalError;
 import com.linbit.linstor.core.identifier.RemoteName;
 import com.linbit.linstor.core.objects.remotes.AbsRemote;
 import com.linbit.linstor.core.objects.remotes.EbsRemote;
 import com.linbit.linstor.core.objects.remotes.S3Remote;
 import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.dbdrivers.interfaces.remotes.EbsRemoteDatabaseDriver;
+import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.ObjectProtectionFactory;
 import com.linbit.linstor.transaction.TransactionObjectFactory;
@@ -23,6 +24,7 @@ import java.util.UUID;
 
 public class EbsRemoteSatelliteFactory
 {
+    private final ErrorReporter errorReporter;
     private final EbsRemoteDatabaseDriver driver;
     private final TransactionObjectFactory transObjFactory;
     private final Provider<TransactionMgr> transMgrProvider;
@@ -31,6 +33,7 @@ public class EbsRemoteSatelliteFactory
 
     @Inject
     public EbsRemoteSatelliteFactory(
+        ErrorReporter errorReporterRef,
         CoreModule.RemoteMap remoteMapRef,
         EbsRemoteDatabaseDriver driverRef,
         ObjectProtectionFactory objectProtectionFactoryRef,
@@ -38,6 +41,7 @@ public class EbsRemoteSatelliteFactory
         Provider<TransactionMgr> transMgrProviderRef
     )
     {
+        errorReporter = errorReporterRef;
         remoteMap = remoteMapRef;
         driver = driverRef;
         objectProtectionFactory = objectProtectionFactoryRef;
@@ -89,7 +93,8 @@ public class EbsRemoteSatelliteFactory
         {
             if (!remote.getUuid().equals(uuid))
             {
-                throw new DivergentUuidsException(
+                CriticalError.dieUuidMissmatch(
+                    errorReporter,
                     S3Remote.class.getSimpleName(),
                     remote.getName().displayValue,
                     remoteNameRef.displayValue,