+++ satellite/src/main/java/com/linbit/linstor/core/objects/S3RemoteSatelliteFactory.java
@@ -3,12 +3,13 @@ package com.linbit.linstor.core.objects;
 import com.linbit.ImplementationError;
 import com.linbit.linstor.core.CoreModule;
 import com.linbit.linstor.core.CoreModule.RemoteMap;
-import com.linbit.linstor.core.DivergentUuidsException;
+import com.linbit.linstor.core.CriticalError;
 import com.linbit.linstor.core.identifier.RemoteName;
 import com.linbit.linstor.core.objects.remotes.AbsRemote;
 import com.linbit.linstor.core.objects.remotes.S3Remote;
 import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.dbdrivers.interfaces.remotes.S3RemoteDatabaseDriver;
+import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.ObjectProtectionFactory;
 import com.linbit.linstor.transaction.TransactionObjectFactory;
@@ -21,6 +22,7 @@ import java.util.UUID;
 
 public class S3RemoteSatelliteFactory
 {
+    private final ErrorReporter errorReporter;
     private final S3RemoteDatabaseDriver driver;
     private final TransactionObjectFactory transObjFactory;
     private final Provider<TransactionMgr> transMgrProvider;
@@ -29,6 +31,7 @@ public class S3RemoteSatelliteFactory
 
     @Inject
     public S3RemoteSatelliteFactory(
+        ErrorReporter errorReporterRef,
         CoreModule.RemoteMap remoteMapRef,
         S3RemoteDatabaseDriver driverRef,
         ObjectProtectionFactory objectProtectionFactoryRef,
@@ -36,6 +39,7 @@ public class S3RemoteSatelliteFactory
         Provider<TransactionMgr> transMgrProviderRef
     )
     {
+        errorReporter = errorReporterRef;
         remoteMap = remoteMapRef;
         driver = driverRef;
         objectProtectionFactory = objectProtectionFactoryRef;
@@ -87,7 +91,8 @@ public class S3RemoteSatelliteFactory
         {
             if (!remote.getUuid().equals(uuid))
             {
-                throw new DivergentUuidsException(
+                CriticalError.dieUuidMissmatch(
+                    errorReporter,
                     S3Remote.class.getSimpleName(),
                     remote.getName().displayValue,
                     remoteNameRef.displayValue,