+++ satellite/src/main/java/com/linbit/linstor/core/objects/StltRemoteSatelliteFactory.java
@@ -4,12 +4,13 @@ import com.linbit.ImplementationError;
 import com.linbit.linstor.annotation.Nullable;
 import com.linbit.linstor.core.CoreModule;
 import com.linbit.linstor.core.CoreModule.RemoteMap;
-import com.linbit.linstor.core.DivergentUuidsException;
+import com.linbit.linstor.core.CriticalError;
 import com.linbit.linstor.core.identifier.RemoteName;
 import com.linbit.linstor.core.objects.remotes.AbsRemote;
 import com.linbit.linstor.core.objects.remotes.StltRemote;
 import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.dbdrivers.noop.NoOpFlagDriver;
+import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.ObjectProtectionFactory;
 import com.linbit.linstor.stateflags.StateFlagsPersistence;
@@ -24,6 +25,7 @@ import java.util.UUID;
 
 public class StltRemoteSatelliteFactory
 {
+    private final ErrorReporter errorReporter;
     private final TransactionObjectFactory transObjFactory;
     private final Provider<TransactionMgr> transMgrProvider;
     private final ObjectProtectionFactory objectProtectionFactory;
@@ -32,12 +34,14 @@ public class StltRemoteSatelliteFactory
 
     @Inject
     public StltRemoteSatelliteFactory(
+        ErrorReporter errorReporterRef,
         CoreModule.RemoteMap remoteMapRef,
         ObjectProtectionFactory objectProtectionFactoryRef,
         TransactionObjectFactory transObjFactoryRef,
         Provider<TransactionMgr> transMgrProviderRef
     )
     {
+        errorReporter = errorReporterRef;
         remoteMap = remoteMapRef;
         objectProtectionFactory = objectProtectionFactoryRef;
         transObjFactory = transObjFactoryRef;
@@ -88,7 +92,8 @@ public class StltRemoteSatelliteFactory
         {
             if (!remote.getUuid().equals(uuid))
             {
-                throw new DivergentUuidsException(
+                CriticalError.dieUuidMissmatch(
+                    errorReporter,
                     StltRemote.class.getSimpleName(),
                     remote.getName().displayValue,
                     remoteNameRef.displayValue,