+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - LVM-Thin snapshots are set to read-only now
+- UUID mismatch will now trigger a System.exit(70) as the best way to recover is to restart the satellite
 
 ## [1.31.2] - 2025-06-11